requirements.txt
Flask==2.2.3
google-cloud-datastore==2.7.1